Reduce registers from u32 to u8

+++ src/bvm/registers.rs
@@ -1,7 +1,7 @@
 use std::cell::RefCell;
 use std::collections::HashMap;
 
-pub struct Registers(RefCell<HashMap<u32, u64>>);
+pub struct Registers(RefCell<HashMap<u8, u64>>);
 
 impl Registers {
     pub fn new() -> Registers {
@@ -11,12 +11,12 @@ impl Registers {
         )
     }
 
-    pub fn exists(&self, register: &u32) -> bool {
+    pub fn exists(&self, register: &u8) -> bool {
         // Check to see if register exists
         self.0.borrow().contains_key(register)
     }
 
-    pub fn get(&self, register: &u32) -> Option<u64> {
+    pub fn get(&self, register: &u8) -> Option<u64> {
         // Get the value stored in register, if exists
         if self.exists(register) {
             Some(self.0.borrow()[register])
@@ -25,7 +25,7 @@ impl Registers {
         }
     }
 
-    pub fn set(&self, register: u32, data: u64) {
+    pub fn set(&self, register: u8, data: u64) {
         // Set the value of a register
         &self.0.borrow_mut().insert(register, data);
     }
